requirements.txt
Django==2.2.*
django-colorfield==0.3.2